+++ pkgs/artisanal/lib/plugins.dart
@@ -59,6 +59,14 @@ export 'src/plugins/remote_surface_notification_service.dart'
     show RemotePluginNotifier, RemotePluginNotificationHostService;
 export 'src/plugins/remote_surface_file_picker_service.dart'
     show RemotePluginFilePickerHandler, RemotePluginFilePickerHostService;
+export 'src/plugins/remote_surface_manifest.dart'
+    show
+        RemotePluginManifestPlacement,
+        RemotePluginManifest,
+        RemotePluginManifestSchemas,
+        RemotePluginManifestValidator,
+        RemotePluginManifestValidationException,
+        loadRemotePluginManifests;
 export 'src/plugins/remote_surface_controller.dart'
     show RemotePluginSurfaceController;
 export 'src/plugins/remote_surface_drawable.dart'